GROOVY-8477: @Immutable-related transformations should be more configurable
